industry_screener/requirements.txt
# 核心依赖
python>=3.10

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据库
SQLAlchemy>=2.0.0
PyMySQL>=1.0.2
pymysql[rsa]

# API相关(iFinD同花顺)
# 注意:iFinD需要单独安装THS SDK,请参考iFinD官方文档
requests>=2.31.0
urllib3>=2.0.0

# 配置管理
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0.0
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# 任务调度
APScheduler>=3.10.0

# 可视化
streamlit>=1.28.0
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 缓存(可选)
redis>=4.5.0

# 工具库
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.65.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# 代码质量
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# 开发工具
ipython>=8.14.0
jupyter>=1.0.0
click>=8.1.0
